use setuptools `console_scripts` entrypoint for pipx
setuptools provides `entry_points.console_scripts`[1] as simple
way to specify command line entry points to python packages.
It generates a "shim" to invoke the code automatically.

One key advantage that the setuptools "shim" provides is that it works
correctly on Windows, installing an executable wrapper that does the
right thing automatically.

pipx[2] extends that to support automatic installation in an isolated
virtual environment, exposing those binaries.  This turns any pypi
module into an "application".

This commit updates the structure of superfsmon to use the
`console_scripts` feature, instead of directly installing a simple
python script.

That makes it practical to install and run on windows — where it works
perfectly correctly, thanks to the watchdog module being portable.

[1]: https://python-packaging.readthedocs.io/en/latest/command-line-scripts.html#the-console-scripts-entry-point
[2]: https://pypa.github.io/pipx/#where-does-pipx-install-apps-from