depend on `supervisor-win` on win32
`supervisor` doesn't run on Windows, only POSIX systems.
The `supervisor-win` package extends it to support win32.

This adjusts the dependencies to use the platform-specific version of
supervisor on Windows.  Since `supervisor-win` installs all of
`supervisor` the existing code works unmodified.

The alternative would be to drop the dependency entirely, which didn't
seem like the right choice.